trivial: Fix integer truncation in AMD GPU PSP firmware parser
Prevent potential parser confusion when processing PSP directory entries
with location values >= 2^32 by using the correct 64-bit type.

(cherry picked from commit 2ee0f252d6e1c69e1fa679a2c53c8e305bbf3231)

+++ plugins/amd-gpu/fu-amd-gpu-psp-firmware.c
@@ -99,7 +99,7 @@ fu_amd_gpu_psp_firmware_parse_l1(FuAmdGpuPspFirmware *self,
 		return FALSE;
 	offset += st_dir->buf->len;
 	for (guint i = 0; i < fu_struct_psp_dir_get_total_entries(st_dir); i++) {
-		guint loc;
+		guint64 loc;
 		guint sz;
 		g_autoptr(FuStructPspDirTable) st_entry = NULL;
 		g_autoptr(FuStructImageSlotHeader) st_hdr = NULL;
